rename HeaderComponent
